DB: add group_users.created_at column
We want to know when a user has been granted a role to a document, workspace or an org.

We introduce the `group_users.created_at column` which stores the creation date of the record
(thus the moment when the role has been set / modified for a user).

We don't expose to the ORM, this column is meant to remain available for queries to external tools.