delete forks owned by a user before deleting that user
Currently, if a user is deleted, forks the owned slip through the
cracks and are reported by the Housekeeper as undeletable. This is
because we lose track of the user id associated with them, and that
happens to be used as part of their composite identifier.

This deletes the fork immediately. There isn't much point holding
onto it if is inaccessible. Deletion of forks needs to be done
somewhat carefully, to coordinate the deletion from a listing in
the database, and the corresponding storage of the document
contents.

To simplify testing of this kind of feature, I ported support of
`DOC_WORKER_COUNT` from `devServerMain` to `MergedServer`.