hold writes on a document if needed, to get a backup made
If SQLite writes are taking a long time, and are happening a lot,
then backups may get delayed. This adds a way for the backup system
to hold writes for a while if it isn't progressing. The threshold
I have for that here may be overly aggressive? In our SaaS, we see
that delayed backups even for very busy documents are already
rare. This change could result in occasional slower editing on
documents that are currently fine.