Avoid using oidc callback URL as post-login URL (#1384)
## Context

Fixes #1383.

## Proposed solution

This PR fixes the referenced issue by making sure the OIDC callback URL
is overwritten in the browser's navigation bar, even when the OIDC login
fails and the final redirection to the desired URL cannot be done.

This is done in the error page using the browser's history API (as used
elsewhere in the Grist client code).

The result is that a failed OIDC login displays the usual error page but
with the current URL set to the originally request URL, so that the user
can either refresh the page for a new attempt, or use the "Sign in
again" button below the error message, which will pick up the originally
requested URL as the new target URL.

This makes the failed OIDC login behaviour consistent with that of other
errors, like attempting to open a document to which the user does not
have access: in such cases, the current URL remains that of the
(inaccessible) document, so that the page can simply be refreshed once
the permission issue is resolved.

## Has this been tested?

<!-- Put an `x` in the box that applies: -->

- [x] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [ ] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->

## Screenshots / Screencasts

<img width="1727" alt="image"
src="https://github.com/user-attachments/assets/477f44c2-8543-4474-83aa-e7e811567daa"
/>

---------

Co-authored-by: Florent <florent.git@zeteo.me>