delete forks owned by a user before deleting that user (#1589)
Currently, if a user is deleted, forks they owned slip through the
cracks and are reported by the Housekeeper as undeletable. This is
because we lose track of the user id associated with the forks, and that
happens to be used as part of the composite identifier for forks,
determining where they are stored on disk and in external storage.

With this change, forks are deleted immediately. There isn't much point
holding onto a fork if it has become inaccessible. Deletion of forks
needs to be done somewhat carefully, to coordinate the deletion from a
listing in the database, and the corresponding storage of the document
contents.

To simplify testing of this kind of feature, I ported support of
`DOC_WORKER_COUNT` from `devServerMain` to `MergedServer`.