DocApi: allow configuring hard limit of MAX_PARALLEL_REQUESTS_PER_DOC
Traditionally this has been set to a hardcoded limit of 10 parallel
requests per doc. This makes sense for the getgrist.com instance, but
not for for some self-hosters. For them, this limit is too
restrictive.

Let us provide an envvar with which this limit may be changed or
removed altogether, while keeping the default at 10.