attempt to make the Importer2 tests more reliable (#1605)
The Importer2 tests have been unreliable for a long time. A few people
(including myself) have looked at it and not been able to pin down
exactly what is up. To be clear, the import behavior looks fine, it is
just something a little erratic in focus or scrolling in the test. This
change introduces a brute force attempt at taming the test. With this
change, the test has been observed to pass 20 times in succession under
github runner conditions.

(see last run of https://github.com/gristlabs/grist-core/pull/1597,
which runs Importer* tests 20 times)