(core) Soft-deleted materials should not grant guest access to workspace/org.
Summary:
Previously, a soft-deleted document shared with a user would still grant that
user guest access to the org with the document (to list it). This, in
particular, prevented the user from deleting themselves. Now, soft-deleted
materials are ignored for the purpose of guest access, and undeleting them
restores guest access.

Includes a number of other changes for the sake of testing.
- Refactors AccountClose test to allow running it without billing enabled (actually this only affects running in dev; Jenkins has been running them in all cases because it's outside the 'billing' suite).
- Fixes BillingPage to show error in subscription summary rather than infinite loader.
- In BillingPage, show 'Delete team site' button in case of error fetching subscription.
- In AccountClose test, limit the need for billingTools to the one test case that
  tests an org with a paid subscription. Refactor other test cases to work without.
- Adds a test case for user deletion with soft-deleted materials.

Test Plan: Includes a new test case for the changed behavior

Reviewers: jarek

Reviewed By: jarek

Differential Revision: https://phab.getgrist.com/D4564