(themes) change the themes prefs logic to prepare for more themes
The `colors` object in the theme user prefs wasn't really used at all.

- We want to know if a theme is of a dark or light variant (the
appearance), and
- we want to know the name of the theme.

But the idea that, at any moment, we want to store both a dark and light
theme info, is not used at all.

Now in the user prefs screen, we choose a theme _name_ (instead of an
appearance), and deduce the theme appearance from it. The theme name is
stored in the existing `colors` object. It's a bit weird but this allows
us to keep the info without changing the schema and having to make
migrations for just this little thing. The comments should help
understand.