(themes) high contrast theme better shows the right panel subtabs
the contrast between items/selected item in the creator panel subtabs is
really light. here is a way to better mark the hovered/selected item
with something different than color only (border width).

I guess this could be something we think about for all themes but at
least for now it's quickly solved for the theme actually needing it
first.