(themes) by default, enable themes even when customCss is used
It's a rather big impact to totally disable theme choice when custom css
is enabled:

- custom css can be used to just change the logo. In that case it's a
bit of a shame it prevents the user from chosing the dark theme
- soon, there should be a high contrast theme added to the list of
themes. again, if using custom css, it's now even an even more
frustrating choice that is removed to the user, if they actually need
better contrast ratios to use the app on an instance that uses custom
css.

So by default now, themes are enabled even with custom css applied.

To support previous behavior, we can pass `themes` in the
`GRIST_HIDE_UI_ELEMENTS` env var. This is great for people having rather
impactful custom css and who don't have the time to update it if
upgrading their grist version. They can just add the env var in the
meantime.