(themes) keep themeName info in theme state
before, the `colors` object could either store the theme name or
directly the theme tokens. I'll see if tests break by removing this
behavior but it seems like, in the prefs, its always the theme name that
is stored?

Anyway, make sure to always keep the theme name info, because it is
useful to know