(core) [grist-saas] bring back version information during prod deploys
Summary:
This adds a small utility to scan logs for Grist version information, tweaks the migration step to print it out reliably, and has the Jenkins job read it. The version used to get acquired from tests, but the tests run separately now.

The bash scripting has a duct-tape and chewing gum quality to it, indeed the whole approach has, but it is relevant only to one job of our internal CI.

Test Plan: manual

Reviewers: georgegevoian

Reviewed By: georgegevoian

Differential Revision: https://phab.getgrist.com/D4565