(core) Showing other users with doc access on the user manager dialog.
Summary:
The `/doc/:did/access` endpoint reveals more information that is presented to a user
on the user manager dialog. For editors and org/ws owners it also returns other users
on the doc or on the resource, who are not shown on the dialog. Now additional users
with document access are shown in the dialog for document editors.

Note: Guests and other team members without document access are still not shown in the
UserManager dialog. They are revealed only in Comments UI (mentions) or if a user
is document owner, they are suggested in auto-complate dropdown.

Test Plan: Updated

Reviewers: georgegevoian

Reviewed By: georgegevoian

Differential Revision: https://phab.getgrist.com/D4582