(themes) fix loading screen CSS flash on dark themes
when loading grist with a dark appearance, we add a couple style rules
that force a dark background, in order to prevent a flash of light >
dark color.

this was actually broke since my updates on themes a few weeks ago
because of a conflict in html IDs. Because the dark-theme specific
bacground snippet gets overriden when the theme has loaded, and there is
a tiny bit of time where the light pattern img is shown before DOM
content is rendered.

This cleans everything ups:

- there is no more ID conflict (the snippet is appended with
`grist-theme-bg` id)
- the CSS added by the snippet is in the `grist-theme` css @layer, to
eases up custom CSS changes
- appearance/theme info is added early as html attributes to help custom
CSS users modify the loading screen as they need
- theme name, in addition to appearance, is stored in localstorage and
is used as default fallback when trying to load current theme, before
checking system prefs.