clipboard: no need for the copy/cut/paste commands check
This was a bit overkill and even something that made stuff break. The
assumption that the clipboard was needed only when copy/cut/paste
commands were registered was false. This made the app behave weirdly in
some cases.

One use case where we understand the assumption is not great: in a grist
doc, switch from a grid view to a chart view, click on the empty space
of a panel, then go back to a grid view. The chart view has no
copy/cut/paste commands registered, so the focus is set back to body
when clicking the empty space. Then going back to a grid view and using
the kb directly doesn't trigger necessarily the `allowFocus` function.
This makes the `input` command break.

It's actually not even needed because the RegionFocusSwitcher adds the
clipboard_group_focus as needed.