Internationalize Chart Widget (#1622)
## Context

https://community.getgrist.com/t/add-a-weblate-component-for-the-chart-widget/9661

The Chart widget is one of the most important widgets, and it is still
not internationalized, which hampers and discourages users, especially
when deploying the tool in large organizations and when providing change
management support.

This PR prepares the Chart Widget component for internationalization by
extracting all hardcoded text strings and replacing them with calls to
the translation function t(). The corresponding English translations
have been added to the en.client.json file to maintain the current
functionality without any visible changes to users.

## Proposed solution

- Replaced all hardcoded UI strings in ChartView.ts with t() function
calls
- Added all extracted strings to static/locales/en.client.json

## Has this been tested?

- [ ] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [X] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->