(core) support banners for attachment usage limits
Summary:
This adds some of the attachment limits as described in:
https://grist.quip.com/FgtaAW5aVlKw/External-attachments-on-SaaS#temp:C:VaL14e275a88ea14281acbbacc74

Here are the `Features` involved:
 * There is an existing `baseMaxAttachmentsBytesPerDocument` setting.
 * Added a `maxAttachmentsBytesPerOrg` setting.
 * Added a placeholder for a `installationMaxAttachmentsBytes` setting.

No distinction is made between internal and external attachment usage, for the purposes of these limits. In practice internal attachments cause performance issues, and it would be good to have a clear nudge on documents with a lot of internal attachments, but this isn't covered here.

Limits on attachment usage at the individual document level is supported by the same mechanism as row counts and overall file size.

The site-level banner that until now triggers when some individual documents exceed limits will now also trigger when the site-level attachment limit is hit. This limit does not yet have teeth.

Installation level attachment limits are not enforced yet.

Test Plan: added a test, updated existing tests

Reviewers: jarek

Reviewed By: jarek

Differential Revision: https://phab.getgrist.com/D4555