(kb) trap focus inside panels
this makes sure that when we tab out of bounds of a panel, we wrap to
the first (or last) element in it instead of going out of it.

this is to prevent focusing back the active section when tabbing out of
the panel (since tabbing out of the panel makes document.activeElement
an element not allowed by the clipboard, it focuses back the active
section automatically).