add a test for pyodide (#1656)
Pyodide sandboxing was broken recently when a package version changed
during the removal of Python 2. This adds a small test to verify that
pyodide sandboxing is still working. I have since built and added the
needed package.

The packaged involved is `six`, which in principle may no longer be
needed, but as long as it is listed in `sandbox/requirements.txt` then
the Grist pyodide sandbox is going to need it too.