Makes document attachment usage consider external attachments (#1600)
## Context

Documents' attachment usage currently only considers files which are
stored internally.

If a document has external attachments, they won't show in the
document's "Size of attachments" bar in "Raw Data".

## This commit

This updates the attachment usage calculations to use
`_grist_Attachments.fileSize` for external attachments, while preserving
the existing logic for internal attachments.

#### Missing attachments

If a document has external attachments and is uploaded to a Grist
instance, these missing attachments will still count towards the
document's attachment usage. There's no easy way to mitigate this in all
situations without checking all attachments on document open.

#### Untrusted file size value
This creates a problem with imported documents, which may have an
incorrect `fileSize` value if it has been manually altered (i.e - the
value is untrusted).

To mitigate this, the `fileSize` value is now updated when attachments
are uploaded as a .tar archive. This means that eventually, `fileSize`
is always valid for files which are present in external storage.

## Related issues

#1021 

## Has this been tested?

- [X] 👍 yes, I added tests to the test suite