Update external attachment store when importing a document (#1629)
## Context

When importing a document that uses external attachments from another
installation, the document's attachment store id still refers to the one
on the original installation.

This can cause issues such as:
- API errors when trying to access an invalid store
- Documents being set to 'external' on installations with no external
storage.

The current fix is to toggle the document from 'External' -> 'Internal'
-> 'External' to refresh the store id, but this is opaque and
undocumented, and only works on installations with external stores
configured.

Multiple solutions were considered for this:
* _Add a 'magic value' store id that always refers to the installation's
default store_ - Backwards compatibility would be an issue, with this
problem persisting for documents currently in use.
* _Update the attachment store when a document is opened_ - This could
theoretically lead to unexpected behaviour with document's attachment
store ids changing without notifying the user. Especially in the future
if stores are explicitly chosen for a document.
* _Modify the document on import_ - This avoids issues with surprising
store ID changes for existing documents (as fewer assumptions exist
around doc behaviour during import), but is trickier as we don't open
the document as an ActiveDoc, so existing APIs for modifying the setting
won't work.

## This commit

This commit adds the 'modify document on import' solution.

When a document is imported, the document's settings checked to see if
the attachment store id exists. If not, the attachment store id is
updated to the installation's default store id, or internal if the
installation has no external stores.

## Related issues

#1021