DB: add group_users created_at column (#1592)
## Context

We want to know when a user has been granted a role to a document,
workspace or an org.

## Proposed solution

We introduce a `created_at` column on the `group_users` which stores the
creation date of the record (thus the moment when the role has been set
/ modified for a user).

We don't expose to the ORM, this property is meant to remain available
for queries to external tools.

## Related issues

<!-- If suggesting a new feature or change, please discuss it in an
issue first -->
<!-- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!-- If this does not solve entirely the issue, make also a checklist of
what is done or not: -->

## Has this been tested?

<!-- Put an `x` in the box that applies: -->

- [ ] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [x] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->