choice tokens: only automatically change the text color
in order to prevent background colors drastically changing and making it
maybe weird (the user might think it's a bug, and it might make visually
weird columns) ; we only automatically change colors when bg is set but
text is not. Which is, we assume, the more common scenario.

Compared to before, we check more gray shades to have more chances at
rendering actually readbable tokens. Before, we had only one light and
one dark shade, making it impossible to render readable tokens with some
colors.