AdminPanel: rely on server for state and checks instead of frontend
This discards most of the client-side state on when to check and the
result of the latest check and instead relies on querying the backend
for this state. This includes the result of the latest check as well
as tracking the configuration item whether checking should be done or
not.

As a result, the frontend code has been simplified slightly.