(core) Implementing mentions in comments.
Summary:
Comment's editor allows now to mention user in the comment's text. Editor was rewritten
with editable div to show mentions in a dedicated UI. List of users is taken from /doc/access
endpoint that already contains list of users with document access.

Mentions are stored as comment payloads. The mention itself is just userRef, but the content of
the comment also contains user's name.

Test Plan: Added and updated existing

Reviewers: georgegevoian, dsagal

Reviewed By: georgegevoian, dsagal

Subscribers: dsagal

Differential Revision: https://phab.getgrist.com/D4584