(core) add nudges to use external attachments
Summary:
If a document has more than `GRIST_ATTACHMENT_THRESHOLD_MB` worth of attachments, and they are stored internally, and external storage is available, and the user is an owner, we show a dismissible banner nudging use of external storage and showing where to set it.

An information section with the same text is shown in the Raw Data usage area (not dismissible).

Design: https://grist.quip.com/DOzvAaq8QT7G/Nudges-for-using-external-attachments

Test Plan: added test

Reviewers: jordigh

Reviewed By: jordigh

Differential Revision: https://phab.getgrist.com/D4617