Make distinct variables for logging as JSON or logging api call details
In order to log as JSON, but also to log api details, there used to exist a single variable whose name was not very eloquent (`GRIST_HOSTED_VERSION`).

We propose two distinct variables that we document:
 - `GRIST_LOG_AS_JSON`
 - `GRIST_LOG_API_DETAILS`

And `GRIST_HOSTED_VERSION` which remains not documented and should be considered as deprecated, keep the two behaviors.

⚠️ We introduce a little breaking change, that should not impact administrators: now `GRIST_HOSTED_VERSION=false` is considered deactivating false these behaviors (it used to be considered as a truthy value…)