ToggleEnterpriseModel: don't try to set the busy observable after a page reload
Before this change, in the happy path we'd reload the page but then
attempt to set the busy observable to false. That produced a stack
trace deep in the GrainJS library to the effect of

  TypeError: can't access property "callback", t is null

Instead, make sure that the page reload is done last, right before the
page is reloaded. If for some reason the page wasn't reloaded, set the
busy observable to false too.