ToggleEnterpriseModel: use healthcheck endpoint instead of configAPI endpoint
The config API endpoint was working but sort of accidentally. It was
slow enough to work in most cases.

Instead, we'll use the healthcheck endpoint with the "ready=1" query
param that explicitly checks that the FlexServer has gotten to the end
of its initialisation routine, or, if the server is in the process is
being restarted, is no longer in a ready state.