(core) Fixing bug with saving user preference on public documents
Summary:
On publicly shared documents existing users who are not a member of a team site couldn't
save their preferences as code was checking for org access, even though the org was not
updated or even read.

Now the Grist treats those users as anonymous when working with user prefs.

Test Plan: Added test

Reviewers: Spoffy

Reviewed By: Spoffy

Subscribers: Spoffy

Differential Revision: https://phab.getgrist.com/D4629