(nav) in left panel, make workspace options menus keyboard focusable
the "…" menu of each workspace was only attainable with the mouse on
hover. Now we correctly show the dots also on keyboard focus ; and make
sure they can be focused themselves (by using a button tag).

we also add an alt text on the dots menu to help screen reader users