(a11y) add html region tags to ease up screen reader nav
one way to navigate with a screen reader is by going from html region to
html region. Most common region elements are header, nav, main, footer.
This adds a few region elements on specific parts of the app.

This also changes the role of some of our Region (our way of calling the
main app sections that we can cycle through with kb). The main region is
now a correct main role, and the top panel is now a header.