(nav) workspace items in left panel are more screen reader friendly
in a recent previous commit we made it possible to use keyboard to
target the "…" menu of each workspace in the left panel.

this was great but didn't really take care of screen reader users. The
`<a>` wrapping the `<button>` caused issues.

The html structure was reworked so that the whole row is not an html
`<a>` tag anymore.

Now the row is a div, containing a `<a>` (link to workspace), and a
`<button>` (dots menu) next to each other. The `<a>` tag is still
clickable on the whole row thanks to "stretched link" css.