(kb) trying out something, this needs a followup
the idea is to fix all the interactions where we start focused on the
creator panel and the app ends up focusing the clipboard element by
itself, for example when transforming a column type (after applying the
transform).

In those cases for now we end up in a weird state where the
regionFocusSwitcher says we are on the creator panel, but the
document.activeElement is the hidden textarea (clipboard el).