New variables to log as json or output api details (#1683)
## Context

In order to log as JSON, but also to log api details, there used to
exist a single variable whose name was not very eloquent
(`GRIST_HOSTED_VERSION`).

## Proposed solution

We propose two distinct variables that we document:
 - `GRIST_LOG_AS_JSON`
 - `GRIST_LOG_API_DETAILS`

And `GRIST_HOSTED_VERSION` which remains not documented and should be
considered as deprecated, keep the two behaviors.

⚠️ We introduce a little breaking change, that should not impact
administrators: now `GRIST_HOSTED_VERSION=false` is considered
deactivating false these behaviors (it used to be considered as a truthy
value…)

There are two commits:
 - The first one is the patch itself;
 - The last one is meant to realign the columns;

In the first commit, I took the liberty to change the way we export the
`app/server/lib/log` module in a more ESM-ish fashion. It has few little
impacts in other parts of the code, I hope that don't have much impact.