Duplicate widgets (#1670)
## Context

If a user wants to duplicate a widget, the only option currently is to
do it manually.

## This commit

This adds a 'duplicate widget' option to the widget, allowing the widget
to be copied to any page.

It reuses the duplicatePage logic for the most part, with some changes
to facilitate it working with individual widgets.

## Related issues

#1644