SCIM Groups: Fix migration tests for postgresql
After having introduced the "type" column in the "Groups" table,
the test for migrating TeamMembers failed because the Group entity
had a property that does not exist yet as a column in the table.

Instead execute a plain INSERT INTO statement to run the migration so it
does not involve the non-existing yet "type" column.