(forms) show in the submit button when submission is slow
this is related to attachments that can take a long time to be analyzed
with antivirus software. When form submission takes too long, we show a
spinner + "submitting…" text in the form submission button.

note: we transformed the `disabled` buttons to `aria-disabled` so that
keyboard focus isn't lost and screen readers can still announce the
button content when submitting.