an optimization for querying docs with postgres
In the Grist Labs SaaS, we are seeing the postgres planner having
trouble with the docs query in a large db. There is an X OR Y where
clause that is fast with just X or just Y but combined is much much
slower. Replacing the query with a UNION of two would work but make
an already long query twice as long. Instead this commit adds a
prefiltering CTE that greatly cuts down the number of rows considered,
and makes the query much snappier.

The planning time for these queries has become significant, which could
also be addressed e.g. with prepared statements, but that's left as a
separate issue.