Allow admin to customize any URLs in commonUrls
The grist gouv team often has the need to customize the custom URLs. Thus we open PRs each time.

This can be a bit tedious, long before we have it in production and we
already have plenty URL customisations.

This PR introduces the `GRIST_CUSTOM_COMMON_URLS` env variable, which is expected to be a valid JSON object.

Each keys are documented in the interface descriptor that is moved in a separate module.