an optimization for querying docs with postgres (#1725)
In the Grist Labs SaaS, we are seeing the postgres planner having
trouble with the docs query in a large db. There is an X OR Y where
clause that is fast with just X or just Y but combined is much much
slower. Replacing the query with a UNION of two would work but make an
already long query twice as long. Instead this commit adds a
prefiltering CTE that greatly cuts down the number of rows considered,
and makes the query much snappier.

The where clause is left unchanged because there's a nuance around
aliases. In the where clause, we filter for aliases that exist within
the scope of the orgs the user has access to. In the CTE, we don't, so
it is a weaker filter - plenty strong for efficiency, but if people were
using the ability to set custom urlIds (nobody is though!) the results
are not the same.

The planning time for these queries has become significant, which could
also be addressed e.g. with prepared statements, but that's left as a
separate issue.