add an optional install-level limit for internal attachment storage (#1729)
Add support for `GRIST_MAX_INTERNAL_ATTACHMENTS_BYTES` which, if set,
will cause Grist to refuse to add new attachments if the the total
attachment size would exceed this amount, and the current document is
not set to use external attachments.

Technically Grist docs can have attachments in several stores. This
limit ignores that nuance.