(switches) new switch component ui design
This changes the common switch component code to have switch components
that are hopefully a little bit more understandable:

- when switch if off, the switch pill is on the left, filled, with a
light background and marked borders to correctly see the component (in
gray)
- when switch is on, switch pill is on the right, with a white bg, with
a tick icon inside. the switch bg is filled in accent color.

the previous design had small issues:

- when switch was off, the gray shades used were not really contrasted
with the background
- the design didnt change besides bg color and switch pill moving from
left to right; we can argue these changes were not enough to understand
what is "off" and what is "on".
- in dark mode, design stayed the same instead of adapting its colors

Now, the on and off design differences are more marked, making it less
of a question whether or not the switch is ON or OFF at first sight