update endpoints for deleting sites and workspaces
The easiest-to-find endpoints for deleting sites and workspaces
were old, from an early Grist prototype. The correct method for
deleting sites in particular was hard to find, hidden in a `Doom`
class. This updates the endpoints to use the preferred implementation.

Note that these endpoints remain very unforgiving. Call them to
delete a resource, and it ends up deleted, with no soft-delete
step. There is a /remove endpoint for workspaces (and documents)
that is more forgiving.