[pyodide] move package preparation into a separate script
The `pipe.js` script is used by Grist when running pyodide. It loads
`packages.js` which has functions to help prepare pyodide packages,
but at run-time only needs the ability to list them. So we separate
that out into a separate utility. This helps avoid a desktop app
packaging problem involving dependencies of `node-fetch`:
  https://github.com/gristlabs/grist-desktop/issues/79

There are certainly other ways to resolve the packaging problem, but
this is a particularly simple one.