(refactor) extract the isFocusable function from trapTabKey
trapTabKey internally uses a really helpful function to determine if a
given element is actually focusable or not (by checking not only the tag
or attributes used on the el, but also if it's visible or not).

this will be useful to have as a stand-alone util rather than it being
kinda hidden in the trapTabKey code.