(kb) fix region focus switcher bug when restoring focus on hidden el
There was a bug in the RFS code. Let's say:

- you focus the top panel with Ctrl+O
- you tab through items and end up focused on a button, lets say "share"
- you switch current region by clicking the main widget
- the "share" button gets hidden for whatever reason. Still in the dom,
but invisible
- you use Ctrl+O again to go back on the top panel

There, there was an error: we tried to focus back an unfocusable element
because it was no longer visible.

Now we correctly check if we can actually focus the element before
focusing it, otherwise we focus back the panel element itself.