(core) add GRIST_POSTGRES_USE_PREPARED_STATEMENTS option
Summary:
This adds an optimization to cache plans for queries when using
postgres. It only kicks in for longer queries (in the sense of a
lot of characters in the SQL). We're seeing planning times creeping
up to ~20ms for some queries. The optimization is off by default
and should be considered experimental. It uses prepared statements
to reuse planning effort by caching plans on the database server.

Named prepared statements are scoped to individual connections, so
for each new connection there will be some planning time for each
kind of query. Only subsequent queries on the same connection will
get a boost.

In a test doing 50000 calls to HomeDbManager.getDoc, with the doc
auth cache off, total time with this change was 32 seconds, and
without this change was 5 minutes 18 seconds. In real life, we do
have 5 seconds of caching on that call, and typical individual page
rendering uses just a few queries anyway, so the optimization will
perhaps not be very perceptible. But it hopefully resets somewhat
some of the flab that was building up.

Test Plan: manual, plus basic test

Reviewers: jordigh

Reviewed By: jordigh

Differential Revision: https://phab.getgrist.com/D4674