(tests) fix left panel test logic
It seems the left panel testing auto-expansion had an issue. If we
reload the page without moving the cursor, the left panel will auto
expand again, making the test believe the auto-expansion opening has
been saved to session.

Make sure to move away the mouse so that we correctly test if the left
panel stayed open or not after reload