(tests) wait for potential race condition in version update banner
This test regularly fails in the github CI but I'm having real trouble
reproducing this failing test locally.
It seems to be related to the fakeServer payload not necessarily being
always ready as soon as we think…

"Fixing" this by waiting a tiny bit. Yep, I don't know what I'm doing.