Generate commonUrls server-side
Before this commit, both the clients and the server had to generate the
commonUrls.

This meant that to do so, the env variables to customize them had to be
shared, and they were many of them.

It's cleaner to just let the server compute the common urls and pass
them to the client.