(aria tabs) prevent tabs from collapsing with Clipboard/RegionFocusLayer
By default, elements having a tabindex attribute are set aside by the
clipboard, meaning focus stays on the element if it gets focused.
Also, the RegionFocusSwitcher checks clicks on regions and focus the
region if it detects a tabindex, following the clipboard logic.

This can get in the way when precisely handling keyboard focus on
specific UI parts.

This was the case on the right panel tabs now that they are handled
correctly with keyboard. While we want to keep focus on them when
focused in the region (via Ctrl+O kb shortcut) (the RegionFocusSwitcher
takes care of that for us), we don't want to keep/set focus when just
using the mouse. But with the usage of tabindex, it was always the case,
kb or mouse.

Now there is a new escape hatch utility class to prevent the clipboard
and RFS from autodeciding, preventing the tabs to take focus while using
the mouse.