(mouse) fix region focus switcher bug when clicking regions
when already focused in a region, then clicking something that we
consider "not focusable", we previously did nothing.
This could lead to surprise users:
- i click on a input in the right panel, type in it
- i then click directly on a button in the right panel
- if I press tab, tab is in the right panel. I expect tab to stay in the
active section widget.

To prevent these surprises, now we make sure to always default to
focusing the active section on click, when not clicking on something
actually focusable.