(tiny) set DEFAULT_MAX_ROWS according to deploymentType (#1723)
## Context

For self-hosters, the 20,000 row limit shown in Raw Data is both false
and confusing. A default of 150,000 seems to be more sensible on a
configuration that's aligned with [what the documentation
suggests](https://support.getgrist.com/self-managed/#what-are-the-hardware-requirements-for-hosting-grist).

## Proposed solution

Check the `deploymentType` variable in the Grist config, set
DEFAULT_MAX_ROWS to 20,000 when on the SaaS and to 150,000 otherwise.

I know that in theory you'd just need to rewrite the default plan in the
Features table of the home db in order to set this value, but a sensible
default goes a long way imo.

## Has this been tested?

<!-- Put an `x` in the box that applies: -->

Tested locally and it works.