(kb) RegionFocusSwitcher: dont bother w/ activeSectionId when not needed
Before, we manually set focus to the activeSectionId when we wanted to
go back to the view layout.

This led to issues in the Importer tests suite, where working on a
table, then undoing work via the undo button would change the active
section to a previous state being an entirely different sectionId. In
that case, the RFS would set an obsolete id when using
gristDoc.viewModel.activeSectionId().

I don't really get why I had this behavior… this commit is kind of a
trick to go around the issue. Now we stop manually setting the focus
section to the currently "activeSectionId", we only set back the focus
to the view layout, that itself correctly knows which is the
activeSectionId.

The code could benefit from some cleaning to better handle that. Will
certainly end up refactoring things a bit when things are more clear.