(core) Add assistant endpoint for creating docs
Summary:
A new endpoint, `/assistant/start`, is added to the home server by version 2
of the OpenAI assistant. The endpoint accepts a prompt, and optionally, the ID
of a document to copy, and redirects you to a new document with the assistant
open by default, and the prompt auto-submitted.

The signup pages now supports a `assistantPrompt` query parameter that will
automatically call the new endpoint on the first visit after a signup. The page
also has a new sidebar when the parameter is set, to indicate this new flow.

Test Plan: Browser tests

Reviewers: paulfitz

Reviewed By: paulfitz

Differential Revision: https://phab.getgrist.com/D4670